requirements.txt
python-bittrex==0.1.4
